fix(proc): widen stackObj pointer arithmetic to uintptr_t in PROC_BirthWithObject
th->object = (void *)(((u32)stackObj) + 8) truncated the freshly-allocated
stack/object pointer to 32 bits before reconstructing it -- the same
function's own PROC_DestroyObject call sites three lines above already use
(uintptr_t)stackObj correctly, this one assignment was just missed by an
earlier truncation sweep. Every thread's th->object (Driver, Instance, etc.)
came out corrupted, crashing the moment anything dereferenced it -- first
hit in VehBirth_Player's memset on the player's own Driver object.

+++ game/PROC.c
@@ -265,7 +265,7 @@ struct Thread *PROC_BirthWithObject(int flags, void *funcThTick, const char *nam
 	// set remaining fields AFTER linking (ASM order)
 	th->funcThTick = funcThTick;
 	th->name = name;
-	th->object = (void *)(((u32)stackObj) + 8);
+	th->object = (void *)(((uintptr_t)stackObj) + 8);
 
 	return th;
 }
